% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocation_kp.R
\name{allocation_KP}
\alias{allocation_KP}
\title{Weight function}
\usage{
allocation_KP(var1, var2, var3, var4, var5)
}
\arguments{
\item{var1}{set the default values.}
}
\description{
This function allows you find a within school allocation.
}
\examples{
allocation_KP()
}
\keyword{allocation}
