% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocation_weights.R
\name{allocation_weights}
\alias{allocation_weights}
\title{Allocation function}
\usage{
allocation_weights(x, y)
}
\arguments{
\item{x}{The data frame gathered will utilized the uploaded data frame.}
}
\description{
This function retrieves the allocation and set the samples.
}
\examples{
allocation_weights()
}
\keyword{allocation}
