% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_df.R
\name{check_df}
\alias{check_df}
\title{Check warnings}
\usage{
check_df(df)
}
\arguments{
\item{df}{The data frame gathered will utilized the uploaded data frame.}
}
\description{
This function check for possible errors in the data.frame.
}
\examples{
check_df()
}
\keyword{warnings}
