% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_df.R
\name{fill_df}
\alias{fill_df}
\title{Fill data frame function}
\usage{
fill_df(x)
}
\arguments{
\item{df}{The data frame gathered will utilized the uploaded data frame.}
}
\description{
This function remove and replace external values with 0.
}
\examples{
fill_df()
}
\keyword{cleaning}
