% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i_startv.R
\name{i_startv}
\alias{i_startv}
\title{Starting values functions}
\usage{
i_startv(x)
}
\arguments{
\item{df}{The data frame gathered will utilized the uploaded data frame.}
}
\description{
This function gathers the number of teachers that work in each ISCED level.
}
\examples{
i_startv()
}
\keyword{values}
