% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_df.R
\name{prepare_df}
\alias{prepare_df}
\title{Prepare dataframe function}
\usage{
prepare_df(x)
}
\arguments{
\item{df}{The data frame gathered will utilized the uploaded data frame.}
}
\description{
This function extract the relevant values from the dataframe.
}
\examples{
prepare_df()
}
\keyword{values}
