% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_case.R
\name{single_case}
\alias{single_case}
\title{Single case}
\usage{
single_case(x, y)
}
\arguments{
\item{x}{The data frame gathered will utilized the uploaded data frame.}
}
\description{
This function retrieves the allocation and set the samples.
}
\examples{
single_case()
}
\keyword{allocation}
