% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_file.R
\name{v_file}
\alias{v_file}
\title{Check valid file}
\usage{
v_file(file)
}
\arguments{
\item{file}{Check if the value uploaded is valid..}
}
\description{
This function check for possible errors in the data.frame.
}
\examples{
v_file()
}
\keyword{valitation}
